Ignore undefined/null children
Fixes fail during rendering of google maps component if a child is a null/undefined component.

My use case is that I have a user position component (CurrentPosition) as a child of the GoogleMap component. But if the user has not allowed geoposition, or if the position is unavailable, CurrentPosition is null and not a component, leading to the render() function failing because there is no .props on child.